refactor: fix TS errors
